Move decodePubkey, decodePrivkey, decodeSignature into decode.ts
